requirements.txt
feedparser>=6.0.0
requests>=2.31.0
anthropic>=0.25.0
python-dateutil>=2.8.0
PyYAML>=6.0
flask>=3.0.0